requirements.txt
aiohttp==3.7.4.post0
beautifulsoup4==4.12.2
colorama==0.4.6
colored==2.2.3
discord.py==2.3.2
discord_webhook==1.2.1
Pillow==10.0.0
Requests==2.31.0